fix:code check

+++ test/unittest/process_dump/dump_info_header_test.cpp
@@ -37,7 +37,7 @@ public:
     static void TearDownTestCase(void) {}
     void SetUp();
     void TearDown() {}
-    static int WriteLogFunc(int32_t fd, const char *buf, int len);
+    static int WriteLogFunc(int32_t fd, const char *buf, size_t len);
     static std::string result;
 };
 } // namespace HiviewDFX
@@ -55,7 +55,7 @@ void DumpInfoHeaderTest::SetUp(void)
     DfxBufferWriter::GetInstance().SetWriteFunc(DumpInfoHeaderTest::WriteLogFunc);
 }
 
-int DumpInfoHeaderTest::WriteLogFunc(int32_t fd, const char *buf, int len)
+int DumpInfoHeaderTest::WriteLogFunc(int32_t fd, const char *buf, size_t len)
 {
     DumpInfoHeaderTest::result.append(std::string(buf, len));
     return 0;
